alter gifts to new stuff

+++ migrations/20160930122947_gifts.js
@@ -5,8 +5,8 @@ exports.up = knex =>
         table.string('image_url').notNullable();
         table.string('name').notNullable();
         table.integer('requested').notNullable();
-        table.decimal('price').notNullable();
-        table.integer('position').notNullable();
+        table.integer('price').notNullable();
+        table.float('position').notNullable();
         table.integer('wedding_profile_id').references('wedding_profiles.id').notNullable();
     });
